requirements.txt
# Core Dependencies
torch==2.7.1
transformers==4.53.3
accelerate==1.8.1

# FastChat for conversation templates
fschat==0.2.36

# Hugging Face Hub
huggingface-hub==0.33.2
safetensors==0.5.3

# Scientific Computing
numpy==2.0.2
scipy==1.13.1

# Visualization
matplotlib==3.9.4

# YAML Configuration
PyYAML==6.0.2

# GPU Monitoring (optional but recommended)
nvidia-ml-py3==7.352.0
pynvml

# Additional utilities
tqdm==4.67.1
